// packages/fhir-r4/src/v4.0.1/capabilitystatementsearchparam/types.ts
import type { Extension, Element } from "../core/types";

/** Generated from FHIR JSON Schema */

/** A Capability Statement documents a set of capabilities (behaviors) of a FHIR Server for a particular version of FHIR that may be used as a statement of actual server functionality or a statement of required or desired server implementation. */
export interface CapabilityStatementSearchParam {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  name?: string;

  _name?: Element;

  definition?: string;

  type?:
    | "number"
    | "date"
    | "string"
    | "token"
    | "reference"
    | "composite"
    | "quantity"
    | "uri"
    | "special";

  _type?: Element;

  documentation?: string;

  _documentation?: Element;
}